# openapi.yaml
...
openapi: 3.0.1
info:
  title: Dataspace Connector
  description: IDS Connector reference implementation
  contact:
    name: Dataspace Connector
    url: https://github.com/International-Data-Spaces-Association/DataspaceConnector
    email: support@sovity.de
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0.txt
  version: 8.0.0
servers:
  - url: https://localhost:8080
    description: Generated server url
security:
  - basicAuth: []
...
paths:
...
  /api/rules/{id}/contracts:
    get:
      tags:
        - Rules
      summary: Get all children of a base resource with pagination.
      operationId: getResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelContractView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelContractView'
    put:
      tags:
        - Rules
      summary: Replace the children of a base resource.
      operationId: replaceResources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Rules
      summary: Add a list of children to a base resource.
      operationId: addResources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelContractView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelContractView'
    delete:
      tags:
        - Rules
      summary: Remove a list of children from a base resource.
      operationId: removeResources
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
  /api/routes/{id}/steps:
    get:
      tags:
        - Routes
      summary: Get all children of a base resource with pagination.
      operationId: getResource_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRouteView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRouteView'
    put:
      tags:
        - Routes
      summary: Replace the children of a base resource.
      operationId: replaceResources_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Routes
      summary: Add a list of children to a base resource.
      operationId: addResources_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRouteView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRouteView'
    delete:
      tags:
        - Routes
      summary: Remove a list of children from a base resource.
      operationId: removeResources_1
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
  /api/requests/{id}/representations:
    get:
      tags:
        - Requested Resources
      summary: Get all children of a base resource with pagination.
      operationId: getResource_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
    put:
      tags:
        - Requested Resources
      summary: Replace the children of a base resource.
      operationId: replaceResources_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Requested Resources
      summary: Add a list of children to a base resource.
      operationId: addResources_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
    delete:
      tags:
        - Requested Resources
      summary: Remove a list of children from a base resource.
      operationId: removeResources_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
  /api/requests/{id}/catalogs:
    get:
      tags:
        - Requested Resources
      summary: Get all children of a base resource with pagination.
      operationId: getResource_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelCatalogView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelCatalogView'
    put:
      tags:
        - Requested Resources
      summary: Replace the children of a base resource.
      operationId: replaceResources_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Requested Resources
      summary: Add a list of children to a base resource.
      operationId: addResources_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelCatalogView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelCatalogView'
    delete:
      tags:
        - Requested Resources
      summary: Remove a list of children from a base resource.
      operationId: removeResources_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
  /api/representations/{id}/requests:
    get:
      tags:
        - Representations
      summary: Get all children of a base resource with pagination.
      operationId: getResource_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRequestedResourceView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRequestedResourceView'
    put:
      tags:
        - Representations
      summary: Replace the children of a base resource.
      operationId: replaceResources_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Representations
      summary: Add a list of children to a base resource.
      operationId: addResources_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRequestedResourceView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRequestedResourceView'
    delete:
      tags:
        - Representations
      summary: Remove a list of children from a base resource.
      operationId: removeResources_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
  /api/representations/{id}/offers:
    get:
      tags:
        - Representations
      summary: Get all children of a base resource with pagination.
      operationId: getResource_8
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelOfferedResourceView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelOfferedResourceView'
    put:
      tags:
        - Representations
      summary: Replace the children of a base resource.
      operationId: replaceResources_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Representations
      summary: Add a list of children to a base resource.
      operationId: addResources_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelOfferedResourceView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelOfferedResourceView'
    delete:
      tags:
        - Representations
      summary: Remove a list of children from a base resource.
      operationId: removeResources_5
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
  /api/representations/{id}/artifacts:
    get:
      tags:
        - Representations
      summary: Get all children of a base resource with pagination.
      operationId: getResource_9
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelArtifactView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelArtifactView'
    put:
      tags:
        - Representations
      summary: Replace the children of a base resource.
      operationId: replaceResources_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Representations
      summary: Add a list of children to a base resource.
      operationId: addResources_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelArtifactView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelArtifactView'
    delete:
      tags:
        - Representations
      summary: Remove a list of children from a base resource.
      operationId: removeResources_6
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
  /api/offers/{id}/representations:
    get:
      tags:
        - Offered Resources
      summary: Get all children of a base resource with pagination.
      operationId: getResource_11
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 30
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
    put:
      tags:
        - Offered Resources
      summary: Replace the children of a base resource.
      operationId: replaceResources_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
    post:
      tags:
        - Offered Resources
      summary: Add a list of children to a base resource.
      operationId: addResources_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "200":
          description: Ok
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagedModelRepresentationView'
    delete:
      tags:
        - Offered Resources
      summary: Remove a list of children from a base resource.
      operationId: removeResources_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
                format: uri
        required: true
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    PagedModelContractView:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            contracts:
              type: array
              items:
                $ref: '#/components/schemas/ContractView'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PagedModelRouteView:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            routes:
              type: array
              items:
                $ref: '#/components/schemas/RouteView'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PagedModelRepresentationView:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            representations:
              type: array
              items:
                $ref: '#/components/schemas/RepresentationView'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PagedModelCatalogView:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            catalogs:
              type: array
              items:
                $ref: '#/components/schemas/CatalogView'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PagedModelRequestedResourceView:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            resources:
              type: array
              items:
                $ref: '#/components/schemas/RequestedResourceView'
        _links:
          $ref: '#/components/schemas/Links'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic